+++ components/updater/card/src/main/java/com/flipperdevices/updater/card/composable/ComposableFirmwareUpdaterError.kt
@@ -20,6 +20,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.flipperdevices.core.ui.ktx.painterResourceByKey
 import com.flipperdevices.core.ui.res.R as DesignSystem
+import com.flipperdevices.core.ui.theme.FlipperThemeInternal
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
 import com.flipperdevices.updater.card.R
@@ -106,9 +107,11 @@ fun getImageByUpdateError(type: UpdateErrorType): Int {
 )
 @Composable
 private fun ComposableFirmwareUpdaterErrorPreview() {
-    Column {
-        ComposableFirmwareUpdaterError(UpdateErrorType.NO_INTERNET)
-        ComposableFirmwareUpdaterError(UpdateErrorType.UNABLE_TO_SERVER)
-        ComposableFirmwareUpdaterError(UpdateErrorType.NO_SD_CARD)
+    FlipperThemeInternal {
+        Column {
+            ComposableFirmwareUpdaterError(UpdateErrorType.NO_INTERNET)
+            ComposableFirmwareUpdaterError(UpdateErrorType.UNABLE_TO_SERVER)
+            ComposableFirmwareUpdaterError(UpdateErrorType.NO_SD_CARD)
+        }
     }
 }