+++ components/updater/card/src/main/java/com/flipperdevices/updater/card/composable/ComposableUpdaterCard.kt
@@ -1,12 +1,18 @@
 package com.flipperdevices.updater.card.composable
 
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.flipperdevices.info.shared.ComposableDeviceInfoRow
-import com.flipperdevices.info.shared.ComposableInfoDivider
+import com.flipperdevices.core.ui.theme.FlipperThemeInternal
 import com.flipperdevices.info.shared.InfoElementCard
 import com.flipperdevices.updater.card.R
 import com.flipperdevices.updater.card.composable.dialogs.ComposableFailedUpdate
@@ -14,9 +20,11 @@ import com.flipperdevices.updater.card.composable.dialogs.ComposableSuccessfulUp
 import com.flipperdevices.updater.card.model.FlipperUpdateState
 import com.flipperdevices.updater.card.viewmodel.UpdateCardViewModel
 import com.flipperdevices.updater.card.viewmodel.UpdateStateViewModel
+import com.flipperdevices.updater.model.DistributionFile
 import com.flipperdevices.updater.model.FirmwareChannel
 import com.flipperdevices.updater.model.FirmwareVersion
 import com.flipperdevices.updater.model.UpdateCardState
+import com.flipperdevices.updater.model.UpdateErrorType
 
 @Composable
 internal fun ComposableUpdaterCardInternal(
@@ -49,53 +57,84 @@ internal fun ComposableUpdaterCardInternal(
         else -> error("Can't find this device status")
     }
 
+    val cardState by updateCardViewModel.getUpdateCardState().collectAsState()
+    ComposableUpdaterCard(
+        modifier = modifier,
+        cardStateLocal = cardState,
+        onSelectChannel = updateCardViewModel::onSelectChannel,
+        retryUpdate = updateCardViewModel::retry
+    )
+}
+
+@Composable
+private fun ComposableUpdaterCard(
+    cardStateLocal: UpdateCardState,
+    modifier: Modifier = Modifier,
+    onSelectChannel: (FirmwareChannel) -> Unit = {},
+    retryUpdate: () -> Unit = {}
+) {
     InfoElementCard(
         modifier = modifier,
         titleId = R.string.updater_card_updater_title
     ) {
-        val cardState by updateCardViewModel.getUpdateCardState().collectAsState()
-        val cardStateLocal = cardState
-
         when (cardStateLocal) {
             is UpdateCardState.Error -> {
                 ComposableFirmwareUpdaterError(
                     typeError = cardStateLocal.type,
-                    retryUpdate = updateCardViewModel::retry
+                    retryUpdate = retryUpdate
                 )
             }
             UpdateCardState.InProgress -> ComposableFirmwareUpdaterContent(
                 version = null,
                 updateCardState = cardStateLocal,
-                onSelectFirmwareChannel = updateCardViewModel::onSelectChannel
+                onSelectFirmwareChannel = onSelectChannel
             )
             is UpdateCardState.NoUpdate -> ComposableFirmwareUpdaterContent(
                 version = cardStateLocal.flipperVersion,
                 updateCardState = cardStateLocal,
-                onSelectFirmwareChannel = updateCardViewModel::onSelectChannel
+                onSelectFirmwareChannel = onSelectChannel
             )
             is UpdateCardState.UpdateAvailable -> ComposableFirmwareUpdaterContent(
                 version = cardStateLocal.lastVersion,
                 updateCardState = cardStateLocal,
-                onSelectFirmwareChannel = updateCardViewModel::onSelectChannel
+                onSelectFirmwareChannel = onSelectChannel
             )
         }
     }
 }
 
+@Preview
 @Composable
-private fun ComposableFirmwareUpdaterContent(
-    version: FirmwareVersion?,
-    updateCardState: UpdateCardState,
-    onSelectFirmwareChannel: (FirmwareChannel) -> Unit
-) {
-    val inProgress = version == null
-    ComposableDeviceInfoRow(titleId = R.string.updater_card_updater_channel, inProgress = false) {
-        ComposableUpdaterFirmwareVersionWithChoice(
-            modifier = it,
-            onSelectFirmwareChannel = onSelectFirmwareChannel,
-            version = version
+private fun ComposableUpdaterCardPreview() {
+    FlipperThemeInternal {
+        val lastVersion = FirmwareVersion(
+            channel = FirmwareChannel.DEV,
+            version = "1.4.5"
         )
+        val cardStates = listOf(
+            UpdateCardState.InProgress,
+            UpdateCardState.NoUpdate(lastVersion),
+            UpdateCardState.UpdateAvailable(
+                lastVersion = lastVersion,
+                fromVersion = lastVersion,
+                updaterDist = DistributionFile(url = "", sha256 = ""),
+                isOtherChannel = false
+            ),
+            UpdateCardState.UpdateAvailable(
+                lastVersion = lastVersion,
+                fromVersion = lastVersion,
+                updaterDist = DistributionFile(url = "", sha256 = ""),
+                isOtherChannel = true
+            ),
+            UpdateCardState.Error(UpdateErrorType.NO_SD_CARD),
+            UpdateCardState.Error(UpdateErrorType.NO_INTERNET),
+            UpdateCardState.Error(UpdateErrorType.UNABLE_TO_SERVER)
+        )
+        Column(Modifier.verticalScroll(rememberScrollState())) {
+            cardStates.forEach {
+                ComposableUpdaterCard(it)
+                Spacer(modifier = Modifier.height(2.dp))
+            }
+        }
     }
-    ComposableInfoDivider()
-    ComposableUpdateButton(updateCardState, inProgress)
 }