+++ components/updater/card/src/main/java/com/flipperdevices/updater/card/composable/ComposableUpdaterReboot.kt
@@ -11,9 +11,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.flipperdevices.core.ui.ktx.animatedDots
 import com.flipperdevices.core.ui.res.R as DesignSystem
+import com.flipperdevices.core.ui.theme.FlipperThemeInternal
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
 import com.flipperdevices.info.shared.InfoElementCard
@@ -49,5 +51,16 @@ fun ComposableUpdaterReboot(modifier: Modifier) {
                 textAlign = TextAlign.Center
             )
         }
+    }
+}
+
+@Preview(
+    showSystemUi = true,
+    showBackground = true
+)
+@Composable
+private fun ComposableUpdaterRebootPreview() {
+    FlipperThemeInternal {
+        ComposableUpdaterReboot(Modifier)
     }
 }